+++ src/pkg/mod.rs
@@ -19,11 +19,13 @@ use crate::proj::BASE_PKGS;
 use crate::repos::cranlike_metadata::{self, repos_get_packages, ArchivedPackage};
 use crate::textfmt::{reflow, wrap, write_field};
 
+mod deps;
 mod manifest;
 
 pub fn sc_pkg(args: &ArgMatches, mainargs: &ArgMatches) -> Result<(), Box<dyn Error>> {
     match args.subcommand() {
         Some(("available", s)) => sc_pkg_available(s, args, mainargs),
+        Some(("deps", s)) => deps::sc_pkg_deps(s, args, mainargs),
         Some(("info", s)) => sc_pkg_info(s, args, mainargs),
         _ => Ok(()), // unreachable
     }